\name{IPPP-package}
\alias{IPPP-package}
\alias{IPPP}
\docType{package}
\title{
\packageTitle{IPPP}
}
\description{
\packageDescription{IPPP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{IPPP}
\packageIndices{IPPP}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{IPPP}

Maintainer: \packageMaintainer{IPPP}
}
\references{
Hohmann, Niklas. "Conditional Densities and Simulations of Inhomogeneous Poisson Point Processes." arXiv 2019. <arXiv:1901.10754>}

\keyword{ package }

