\name{IPPP-package}
\alias{IPPP-package}
\alias{IPPP}
\docType{package}
\title{
\packageTitle{IPPP}
}
\description{
\packageDescription{IPPP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{IPPP}
\packageIndices{IPPP}

For the vignette, use \code{vignette("IPPP")}
}
\author{
\packageAuthor{IPPP}

Maintainer: \packageMaintainer{IPPP}
}
\references{
Hohmann, Niklas. "Conditional Densities and Simulations of Inhomogeneous Poisson Point Processes: The R package "IPPP"" arXiv 2019. <arXiv:1901.10754>
}

\keyword{ package }
\keyword{ Poisson }
\keyword{ Poisson point process}
\keyword{ Point process }
\keyword{ IPPP }

